+++ testng/src/main/java/io/cucumber/testng/TestCaseResultObserver.java
@@ -1,4 +1,4 @@
-package io.cucumber.junit.platform.engine;
+package io.cucumber.testng;
 
 import io.cucumber.core.eventbus.EventBus;
 import io.cucumber.plugin.event.EventHandler;
@@ -10,7 +10,7 @@ import io.cucumber.plugin.event.Status;
 import io.cucumber.plugin.event.TestCaseFinished;
 import io.cucumber.plugin.event.TestStep;
 import io.cucumber.plugin.event.TestStepFinished;
-import org.opentest4j.TestAbortedException;
+import org.testng.SkipException;
 
 import java.net.URI;
 import java.util.ArrayList;
@@ -19,29 +19,33 @@ import java.util.Map;
 import java.util.TreeMap;
 
 import static io.cucumber.plugin.event.Status.PASSED;
+import static io.cucumber.plugin.event.Status.PENDING;
 import static io.cucumber.plugin.event.Status.SKIPPED;
 import static io.cucumber.plugin.event.Status.UNDEFINED;
-import static org.junit.platform.commons.util.ExceptionUtils.throwAsUncheckedException;
 
 class TestCaseResultObserver implements AutoCloseable {
+
+    private static final String SKIPPED_MESSAGE = "This scenario is skipped";
 
     private final EventPublisher bus;
     private final Map<StepLocation, List<String>> snippetsPerStep = new TreeMap<>();
     private final List<UndefinedStepException.Suggestion> suggestions = new ArrayList<>();
     private final EventHandler<SnippetsSuggestedEvent> snippetsSuggested = this::handleSnippetSuggestedEvent;
     private final EventHandler<TestStepFinished> testStepFinished = this::handleTestStepFinished;
+    private final boolean strict;
     private Result result;
     private final EventHandler<TestCaseFinished> testCaseFinished = this::handleTestCaseFinished;
 
-    private TestCaseResultObserver(EventPublisher bus) {
+    private TestCaseResultObserver(EventPublisher bus, boolean strict) {
         this.bus = bus;
+        this.strict = strict;
         bus.registerHandlerFor(SnippetsSuggestedEvent.class, snippetsSuggested);
         bus.registerHandlerFor(TestStepFinished.class, testStepFinished);
         bus.registerHandlerFor(TestCaseFinished.class, testCaseFinished);
     }
 
-    static TestCaseResultObserver observe(EventBus bus) {
-        return new TestCaseResultObserver(bus);
+    static TestCaseResultObserver observe(EventBus bus, boolean strict) {
+        return new TestCaseResultObserver(bus, strict);
     }
 
     @Override
@@ -88,14 +92,20 @@ class TestCaseResultObserver implements AutoCloseable {
         result = event.getResult();
     }
 
-    void assertTestCasePassed() {
+    void assertTestCasePassed() throws Throwable {
         Status status = result.getStatus();
         if (status.is(PASSED)) {
             return;
         }
         Throwable error = result.getError();
-        if (status.is(SKIPPED) && error == null) {
-            throw new TestAbortedException();
+        if (status.is(SKIPPED)) {
+            if (error == null) {
+                throw new SkipException(SKIPPED_MESSAGE);
+            }
+            if (error instanceof SkipException) {
+                throw error;
+            }
+            throw new SkipException(result.getError().getMessage(), error);
         }
 
         if (status.is(UNDEFINED)) {
@@ -103,14 +113,18 @@ class TestCaseResultObserver implements AutoCloseable {
                 // Empty scenario
                 return;
             }
-            throw new UndefinedStepException(suggestions);
+
+            throw new UndefinedStepException(suggestions, strict);
         }
 
-        throw throwAsUncheckedException(error);
+        if (status.is(PENDING) && !strict) {
+            throw new SkipException(result.getError().getMessage(), result.getError());
+        }
+
+        throw error;
     }
 
     private static final class StepLocation implements Comparable<StepLocation> {
-
         private final URI uri;
         private final int line;
 
@@ -125,4 +139,5 @@ class TestCaseResultObserver implements AutoCloseable {
             return order != 0 ? order : Integer.compare(line, o.line);
         }
     }
+
 }